requirements.txt
asgiref==3.3.2
boto3==1.17.53
botocore==1.20.53
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.25.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-appconf==1.0.4
django-filter==2.4.0
django-googledrive-storage==1.5.2
django-storages==1.11.1
django-webpush==0.3.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
http-ece==1.1.0
httplib2==0.19.1
idna==2.10
jmespath==0.10.0
oauth2client==4.1.3
packaging==20.9
Pillow==8.2.0
protobuf==3.15.8
psycopg2==2.8.6
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywebpush==1.9.4
requests==2.25.1
rsa==4.7.2
s3transfer==0.3.7
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
